# api/openapi/valori-v1.yaml
components:
  schemas:
    ApiError:
      description: |-
        The canonical error body, as a schema-bearing DTO.

        [`valori_engine::EngineError`] produces this shape at runtime but lives in
        a crate that does not depend on `utoipa`, so the schema is declared here —
        the translation layer §36 asks for, rather than a re-export of an internal
        type.
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable machine-readable code. Branch on this.
        error:
          description: Human-readable message. Do not parse.
          type: string
      required:
      - error
      - code
      type: object
...
    BuildableIndexKind:
      description: |-
        The index kinds `POST /v1/namespaces/{name}/index` will actually build.

        Phase API-3.3: narrower than the project-wide `IndexKindInput`, and
        deliberately so. The build task in both routers matches on exactly three
        strings — `"hnsw"`, `"ivf"`, `"bq"` — and its `_` arm returns
        `"unknown index type '<x>'"`. `brute` and `auto` are project-level
        selections, not buildable per-collection ANN structures, so they are not
        members here; sending one is an error, and the schema now says so instead
        of advertising an open `string`.

        `null` is the fourth valid value and means *drop the index*. It is carried
        by the `Option`, not by a variant.
      enum:
      - hnsw
      - ivf
      - bq
      type: string
...
    ErrorCode:
      description: |-
        Mirror of [`valori_engine::ErrorCode`] for schema generation.

        `tests/api_contract.rs` diffs the runtime enum against the committed YAML;
        this type exists so the generated document carries the same closed set.
      enum:
      - validation_error
      - unauthorized
      - forbidden
      - not_found
      - collection_not_found
      - record_not_found
      - dimension_mismatch
      - invalid_metric
      - invalid_index
      - index_build_failed
      - conflict
      - capacity_exceeded
      - not_leader
      - unavailable
      - not_implemented
      - internal_error
      type: string
...
    IndexBuildParameters:
      description: |-
        The tuning knobs `POST /v1/namespaces/{name}/index` actually reads.

        Phase API-3.3: [`IndexBuildRequest::parameters`] is a `serde_json::Value`,
        which utoipa rendered as a schema with no `type` at all — `unknown` in
        TypeScript, `Any` in Python, and nothing whatsoever for a user to discover
        the knob names from. It was the only genuinely untyped field in the public
        surface.

        The runtime is not actually open-ended: both routers read exactly five
        keys, all unsigned integers — `m`, `ef_construction`, `ef_search` for HNSW
        (`server.rs` / `cluster_server.rs`, the `"hnsw"` arm) and `n_list`,
        `n_probe` for IVF (the `"ivf"` arm). This type names them.

        `additionalProperties` stays open because the documented behaviour is that
        unknown keys are *ignored*, not rejected — so a client sending one is not
        making an error, and the schema must not claim otherwise.
      properties:
        ef_construction:
          description: 'HNSW: candidate-list size during construction.'
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        ef_search:
          description: 'HNSW: candidate-list size during search.'
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        m:
          description: 'HNSW: neighbours per node. `m_max0` is derived as `2 * m`.'
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        n_list:
          description: 'IVF: centroid count. Omit to auto-scale to `max(16, sqrt(N))`.'
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        n_probe:
          description: 'IVF: probe count. Omit to auto-scale to `max(1, sqrt(n_list))`.'
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
      type: object
    IndexBuildRequest:
      description: The request body for `POST /v1/namespaces/{name}/index`.
      properties:
        parameters:
          $ref: '#/components/schemas/IndexBuildParameters'
          description: |-
            Optional parameter overrides. Only the parameters the implementation
            actually reads are used; unknown keys are ignored.

            Stays a `serde_json::Value` at runtime — the handlers index into it by
            key and tolerate anything — but is *described* as
            [`IndexBuildParameters`] so a generated SDK can offer the real knobs.
        type:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/BuildableIndexKind'
            description: '`"hnsw"`, `"ivf"`, `"bq"`, or `null` (drop the index).'
      type: object
...
    IndexStatusResponse:
      description: |-
        Response to `POST /v1/namespaces/{name}/index` and
        `GET /v1/namespaces/{name}/index`.

        # Cluster vs standalone distinction

        In cluster mode, `desired_type` is always populated from the Raft-
        replicated desired spec (what the cluster wants), while `active_type`
        and `status` reflect this **node's local** build state. They may
        differ temporarily as builds propagate across replicas.

        Example during a transition:
        ```json
        { "desired_type": "ivf", "active_type": "hnsw", "status": "building",
          "building_generation": 2, "active_generation": 1 }
        ```

        In standalone mode, `desired_type` is always equal to `active_type` once
        a build completes (there's only one node).
      properties:
        active_generation:
          description: The active generation number, if any.
          format: int32
          minimum: 0
          type:
          - integer
          - 'null'
        active_type:
          description: The currently serving index type ("hnsw", "ivf", "bq", "none").
          type: string
        base_lsn:
          description: The base LSN of the building generation.
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        build_started_at:
          description: Unix seconds when the current build started.
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        building_generation:
          description: If a build is in progress, its generation number.
          format: int32
          minimum: 0
          type:
          - integer
          - 'null'
        collection:
          type: string
        desired_type:
          description: |-
            The type the user requested (may differ from active while building).
            In cluster mode, this comes from the Raft-replicated desired spec and
            is authoritative for the whole cluster, not just the responding node.
          type:
          - string
          - 'null'
        error:
          description: Human-readable failure reason, if the last build failed.
          type:
          - string
          - 'null'
        status:
          description: Current lifecycle status of the active or building generation.
          type: string
      required:
      - collection
      - active_type
      - status
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Project API key or the legacy VALORI_AUTH_TOKEN bearer token.
      scheme: bearer
      type: http
info:
  description: 'The Valori Data Plane REST API v1, emitted directly from `#[utoipa::path]` annotations on the registered axum handlers. Every public route the Rust routers register appears here and nothing else does — `scripts/verify-api-route-contract.py` proves that equality on every run of the contract gate, and fails on any drift in either direction. Administrative and operator-internal routes (key management, cluster membership, replication streams, `/metrics`) are deliberately excluded: they are served by the node but are not part of the SDK surface.'
  license:
    identifier: MIT OR Apache-2.0
    name: MIT OR Apache-2.0
  title: Valori Data Plane API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/namespaces/{name}/index:
    get:
      description: '`desired_type` is what was asked for; `active_type` and `status` describe what this node is actually serving. In cluster mode `desired_type` comes from the Raft-replicated spec and is cluster-wide, while activation is node-local — the two differ while a build propagates.'
      operationId: get_collection_index
      parameters:
      - description: Collection name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexStatusResponse'
          description: Lifecycle state
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: No such collection
      security:
      - BearerAuth: []
      summary: Read one collection's index lifecycle state
      tags:
      - index
      x-required-scope: read_only
      x-sdk: true
    post:
      description: '`type` is `hnsw`, `ivf`, `bq`, or null to drop the index and revert to exact search. A build is asynchronous: 202 means the build started, and the response carries the building generation. Poll the GET form for completion.'
      operationId: set_collection_index
      parameters:
      - description: Collection name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexBuildRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexStatusResponse'
          description: Index dropped
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexStatusResponse'
          description: Build accepted and started
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unsupported index type
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: No such collection
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: A build is already in progress
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ANN index management unavailable on this node
      security:
      - BearerAuth: []
      summary: Create, change, or drop a collection index
      tags:
      - index
      x-required-scope: read_write
      x-sdk: true
...
servers:
- description: 'The node this document was generated by. Valori is self-hosted: point your client at your own node''s base URL.'
  url: /